# public/openapi.yaml
...
openapi: 3.1.0
info:
  title: Card Orb API
  version: 1.0.0
  summary: A Pokémon card collection, priced, served to a web tool and an iOS app.
  description: |
    Card Orb keeps one person's Pokémon card collection in Postgres, matches it
    against three card catalogues, prices it from Cardmarket, and serves it here.

    **Two clients read this API and both are ours**: the web tool at cardorb.com
    and the iOS app. Nothing here is designed for a third party yet.

    **Every failure is `{ "error": "<sentence>" }`** at a status a client can
    branch on. The sentence is for a person to read; the status is for the code.

    **The version is in the path.** `/v1` changes only by addition: a new field,
    a new route, a new optional query. Anything that removes or renames is `/v2`.
  contact:
    name: Card Orb
    url: https://cardorb.com
servers:
  - url: https://api.cardorb.com
    description: The API on its own host. `/v1/…` here is `/api/v1/…` on cardorb.com.
  - url: https://cardorb.com/api
    description: The same API under the web tool's own origin, which is what the browser uses.
  - url: http://localhost:3000/api
    description: Local development.
...
security:
  - bearer: []
  - session: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A Supabase access token, `Authorization: Bearer <jwt>`. What the iOS app
        sends. Verified locally against the project's signing keys.
    session:
      type: apiKey
      in: cookie
      name: binder_session
      description: |
        The browser's session cookie, set by `POST /v1/session`. What the web
        tool sends. Scoped to cardorb.com, so it never reaches api.cardorb.com —
        the browser calls `/api/v1` on its own origin instead.
    cron:
      type: http
      scheme: bearer
      description: Vercel's `CRON_SECRET`, on `/v1/cron/snapshot` only.
...
  responses:
    BadRequest:
      description: The request was understood and refused. The sentence says why.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Invalid request" }
    Unauthorised:
      description: No viewer. Sign in, or send a bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Sign in to see this." }
    Forbidden:
      description: The request came from an origin this deployment does not allow.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Forbidden" }
    NotFound:
      description: Nothing by that name.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: The body is bigger than this route's `BODY_LIMIT`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Payload too large" }
    UnsupportedMediaType:
      description: A write without `Content-Type application/json`, which is what forces a browser's preflight.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Invalid request" }
    TooManyRequests:
      description: Over this route's rate limit for this address.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Too many requests" }
...
    Unavailable:
      description: This deployment has no database, or the database is not answering.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "This deployment has no database configured." }
  schemas:
    Error:
      type: object
      description: The one shape every failure takes.
      required: [error]
      properties:
        error:
          type: string
          description: A sentence for a person. Branch on the status, not on this.
      additionalProperties: true
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, const: true }
...
    Folder:
      type: object
      description: A folder the person made to sort cards into. The web app's copy calls it a collection.
      required: [id, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        count: { type: integer, description: "Owned copies filed in it. On the list only." }
    FolderDraft:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
...
paths:
...
  /v1/folders/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Collection]
      operationId: renameFolder
      summary: Rename a folder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FolderDraft" }
      responses:
        "200":
          description: Renamed.
          content:
            application/json:
              schema:
                type: object
                required: [ok, folder]
                properties:
                  ok: { type: boolean, const: true }
                  folder: { $ref: "#/components/schemas/Folder" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorised" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/Unavailable" }
    delete:
      tags: [Collection]
      operationId: deleteFolder
      summary: Delete a folder
      description: The copies filed in it stay in the collection and are taken out of the folder first.
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401": { $ref: "#/components/responses/Unauthorised" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/Unavailable" }

  # ── Value ─────────────────────────────────────────────────────────────────
...
  /v1/session:
    post:
      tags: [Account]
      operationId: signIn
      summary: Sign in
      description: |
        Same-origin only; the browser's form. Sets the session cookie. Twenty
        attempts per address and five per address-and-email each fifteen minutes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: That email or password is not right.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Wrong origin, or the address is not confirmed yet (then `unconfirmed` is `true`).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      unconfirmed: { type: boolean, const: true }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/Unavailable" }
    delete:
      tags: [Account]
      operationId: signOut
      summary: Sign out
      description: Clears the session cookie. Always 200, even with no session.
      security: []
      responses:
        "200":
          description: Signed out.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
  /v1/password:
    post:
      tags: [Account]
      operationId: changePassword
      summary: Change the password
      description: Same-origin, session only. `currentPassword` is required unless the caller arrived through recovery.
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 10 }
                currentPassword: { type: string }
      responses:
        "200":
          description: Changed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorised" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/Unavailable" }
...